+++ tkpdcore/src/main/java/com/tokopedia/core/shopinfo/ReputationViewShop.java
@@ -634,15 +634,15 @@ public class ReputationViewShop extends TActivity {
         return new PopupMenu.OnMenuItemClickListener() {
             @Override
             public boolean onMenuItemClick(MenuItem item) {
-                switch (item.getItemId()) {
-                    case R2.id.action_report:
-                        showDialogReport();
-                        return true;
-                    case R2.id.action_delete:
-                        deleteComment();
-                        return true;
-                    default:
-                        return false;
+                int i = item.getItemId();
+                if (i == R.id.action_report) {
+                    showDialogReport();
+                    return true;
+                } else if (i == R.id.action_delete) {
+                    deleteComment();
+                    return true;
+                } else {
+                    return false;
                 }
             }
         };
